+++ Heartland-iOS-SDK/Classes/Extensions/String+Empty.swift
@@ -1,0 +1,12 @@
+//
+//  String+Empty.swift
+//  Heartland-iOS-SDK
+//
+
+import Foundation
+
+public extension String {
+    static var empty: String {
+        return ""
+    }
+}