+++ Heartland-iOS-SDK/Classes/Terminals/PAX/Builders/HpsPaxCreditSaleBuilder.h
@@ -32,7 +32,6 @@
 @property (nonatomic, readwrite) BOOL tipRequest;
 @property (nonatomic, strong) NSString *ecrTransId;
 @property (nonatomic, strong) NSString *clientTransactionId;
-@property (nonatomic, strong) NSString *cardBrandTransactionId;
 
 - (void) execute:(void(^)(HpsPaxCreditResponse*, NSError*))responseBlock;
 - (id)initWithDevice: (HpsPaxDevice*)paxDevice;