+++ Heartland-iOS-SDK/Classes/Terminals/PAX/Builders/HpsPaxCreditVerifyBuilder.h
@@ -24,7 +24,6 @@
 @property (nonatomic, strong) HpsAddress *address;
 @property (nonatomic, readwrite) BOOL requestMultiUseToken;
 @property (nonatomic, strong) NSString *clientTransactionId;
-@property (nonatomic, strong) NSString *cardBrandTransactionId;
 
 - (void) execute:(void(^)(HpsPaxCreditResponse*, NSError*))responseBlock;
 - (id)initWithDevice: (HpsPaxDevice*)paxDevice;