+++ Heartland-iOS-SDK/Classes/Terminals/UPA/Interfaces/HpsUpaParser.m
@@ -77,7 +77,9 @@
         @"BUSY": @(UPA_MSG_TYPE_BUSY),
         @"TO": @(UPA_MSG_TYPE_TIMEOUT),
         @"MSG": @(UPA_MSG_TYPE_MSG),
-        @"DATA": @(UPA_MSG_TYPE_DATA), };
+        @"DATA": @(UPA_MSG_TYPE_DATA),
+//        @"BC": @(UPA_MSG_TYPE_BC),
+    };
     NSNumber *typePointer = upaMessageTypesByRaw[message];
     return typePointer ? [typePointer longValue] : UPA_MSG_TYPE_UNKNOWN;
 }