+++ Heartland-iOS-SDK/Supporting Files/Heartland-iOS-SDK-umbrella.h
@@ -191,6 +191,7 @@
 #import "HpsCardInfoViewController.h"
 #import "XMLDictionary.h"
 #import "HpsAutoSubstantiation.h"
+#import "HpsUpaDeviceSignatureResponse.h"
 
 FOUNDATION_EXPORT double Heartland_iOS_SDKVersionNumber;
 FOUNDATION_EXPORT const unsigned char Heartland_iOS_SDKVersionString[];