+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
  
-## 2.0.8 2023-05-23
+## 2.0.9 2023-06-13
  
 ### Added
-- Added Timeout functionality to PayApp TCP connection. (US344949)
-- Added token and card brand transaction id to the response for Pax. (US345636)
+- Added UPA Signature Data. (US349346)
+- Added UPA Pay App Canada (US346223)
+- Bug fixes and improvements.
  
 ### Changed